requirements.txt
# Core Django Framework
Django==5.2.5

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# CORS and API Documentation
django-cors-headers==4.2.0
drf-yasg==1.21.7

# Caching and Background Tasks
django-redis==5.3.0
celery==5.3.4
redis==5.0.0

# Image Processing
Pillow==11.0.0

# Environment Variables
python-dotenv==1.0.0

# Database (Production)
psycopg2-binary==2.9.9

# Development Dependencies
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Production Dependencies
gunicorn==21.2.0
whitenoise==6.6.0

# Security
django-ratelimit==4.1.0
